+++ app/assets/javascripts/deprecation-silencer/index.js
