+++ app/assets/javascripts/discourse-plugins/package.json
@@ -8,6 +8,7 @@
     "ember-addon"
   ],
   "dependencies": {
+    "deprecation-silencer": "1.0.0",
     "discourse-widget-hbs": "1.0.0",
     "ember-auto-import": "^2.6.3",
     "ember-cli-babel": "^7.26.11",
