+++ app/assets/javascripts/discourse/package.json
@@ -52,6 +52,7 @@
     "bootstrap-json": "1.0.0",
     "broccoli-asset-rev": "^3.0.0",
     "deepmerge": "^4.3.1",
+    "deprecation-silencer": "1.0.0",
     "dialog-holder": "1.0.0",
     "discourse-common": "1.0.0",
     "discourse-plugins": "1.0.0",
